pyproject.toml
[project]
name = "ai-meal-plan-generator"
version = "0.1.0"
description = "AI-powered meal plan generator"
requires-python = ">=3.10"
dependencies = [
    "agent-framework",
    "agent-framework-azure-ai>=1.0.0b260107",
    "agent-framework-devui",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]
